Several improvements
The handling of strings, an especially the occurrence of escaped
quotation marks within strings, has been improved. It results in
individual “char” nodes being reported in the AST, but hopefully
that’s not a problem.

The ambiguity issue that was causing character ranges in sets to be
confused with character strings in sets has been resolved.