Refactored the grammar to fix bugs and support highlighting
The bug that caused a problem with nonterminals named ixml has been
resolved, fix #3.

A number of tokens have been refactored to change what is and isn’t
reported. This is with an eye towards optimizing what you might want
to syntax highlight, not representing the grammar with logical
fidelity.